Update cvp_api.py
Changing back to extend, syntax works as long as input for configlets= variable is a list. Biggest issues with 2018 seem to be the 'id:' kvp not being accepted, and that configlets must be assigned before the topology is saved (one temp session after another is okay as long as save isn't initiated) - If you attempt to assign a configlet after a task has been created to move a device from the undefined container to a regular container, the api gateway will complain that devices in undefined container cannot have configlets assigned to them.

I've added a variable under the deploy_device function to allow the user to select if create_task should be automatically generated during the run, or if they choose to save the topology themselves. True is set as the default so as to match the previous default behavior.